separated message components into own files
