+++ travel-event-host-app/src/components/messaging/message-thread-card/MessageThreadCard.tsx
@@ -1,17 +1,13 @@
+'use client';
 import { UserClient } from '@/app/clients/user/user-client';
 import theme from '@/app/theme';
-import { CustomGenericMuiAvatar } from '@/components/avatar/custom-generic-user-avatar/CustomGenericUserAvatar';
-import { MultiAvatarComponent } from '@/components/avatar/multi-avatar/MultiAvatar';
-import UserAvatar from '@/components/avatar/user-avatar/UserAvatar';
-import { SearchInputField } from '@/components/event-search/search-input-field/SearchInputField';
 import { MessageThread } from '@/models/messaging/message-thread.model';
 import { SecureUser } from '@/types/secure-user';
-import LinearScaleIcon from '@mui/icons-material/LinearScale';
-import LocationPinIcon from '@mui/icons-material/LocationOn';
-import { Box, ButtonBase, Menu, MenuItem, Typography } from '@mui/material';
+import { Box, Typography } from '@mui/material';
 import dayjs from 'dayjs';
 import { useEffect, useState } from 'react';
-import styles from './styles.module.css';
+import { BaseCard } from '../base-card/BaseCard';
+import { SummarySectionMenuActions } from '../summary-section/SummarySection';
 /* 
  Looking at the meetup.com website, the other sections have two variations:
  1. Message thread version:
@@ -43,7 +39,7 @@ export const MessageThreadCard = ({
 }: MessageThreadCardProps) => {
   const [recipients, setRecipients] = useState<Partial<SecureUser>[]>([]);
 
-  const filteredUsers = threadContext.recipients.filter((userId) => userId !== baseUser._id);
+  const filteredUsers = threadContext.recipients.filter((userId) => userId !== baseUser?._id);
   useEffect(() => {
     filteredUsers.map((userId) =>
       UserClient.getUserById(userId).then((user) => setRecipients([...recipients!, user!])),
@@ -67,7 +63,7 @@ export const MessageThreadCard = ({
         {/* TODO: Fix */}
         {/* <SummarySection cardType='thread' user={user} /> */}
         <Box>
-          {/* Blurn of the last message */}
+          {/* Blurb of the last message */}
           <Typography sx={{ color: theme.palette.primary.greyDisabled, fontWeight: 'light' }}>
             {threadContext.messages[threadContext.messages.length - 1].body}
           </Typography>
@@ -82,202 +78,6 @@ export const MessageThreadCard = ({
 const renderMultipleRecipientNames = (users: Partial<SecureUser>[]): string => {
   return users.map((user) => `${user.firstName}`).join(', ');
 };
-
-export const AvatarMessageHeaderCard = ({
-  user,
-  onMenuItemClick,
-}: {
-  user: Partial<SecureUser>;
-  onMenuItemClick: (e: any) => void;
-}) => {
-  return (
-    <BaseCard users={user}>
-      <Box ml={2} width='100%'>
-        <Box>
-          {/* TODO: This needs to be styled differently if it's a thread info */}
-          <Typography
-            sx={{ fontWeight: 'bold', color: theme.palette.primary.charcoal, textAlign: 'left' }}
-          >
-            {/* TODO: Fix */}
-            {user?.firstName} {user?.lastName}
-          </Typography>
-        </Box>
-        {/* TODO: Fix */}
-        <SummarySection cardType='info' user={user} onMenuClick={onMenuItemClick} />
-        <Box>{/* This section is for time ago or some context menu */}</Box>
-      </Box>
-    </BaseCard>
-  );
-};
-
-const BaseCard = ({
-  users,
-  children,
-  backgroundColor,
-  reverseFlow,
-  onCardClicked,
-}: {
-  users: Partial<SecureUser>[] | Partial<SecureUser>;
-  children?: JSX.Element;
-  backgroundColor?: string;
-  reverseFlow?: boolean;
-  onCardClicked?: () => void;
-}) => {
-  // The avatar needs to appear on the left or right
-  return (
-    <Box
-      display='flex'
-      bgcolor={backgroundColor || 'white'}
-      padding={2}
-      flexDirection={reverseFlow ? 'row-reverse' : 'row'}
-      onClick={() => onCardClicked && onCardClicked()}
-      sx={{
-        '&:hover': {
-          backgroundColor: theme.palette.primary.thirdColorIceLight,
-          cursor: 'pointer',
-        },
-      }}
-    >
-      <Box className='avatarEnclosure'>
-        {!Array.isArray(users) ? (
-          <UserAvatar
-            user={users as Partial<SecureUser>}
-            imageClassName={styles.mesageCardAvatar}
-            MuiAvatarComponent={<CustomGenericMuiAvatar theme={theme} />}
-          />
-        ) : (
-          <MultiAvatarComponent users={users} />
-        )}
-      </Box>
-      {children}
-    </Box>
-  );
-};
-
-export interface SummarySectionProps {
-  cardType: 'thread' | 'info';
-  user: Partial<SecureUser>;
-  onMenuClick?: (action: {
-    type: SummarySectionMenuActions;
-    context: { id: string; other?: any };
-  }) => void;
-}
-
-enum SummarySectionMenuActions {
-  RemoveUser = 'removeUser',
-}
-
-const SummarySection = ({ cardType, user, onMenuClick }: SummarySectionProps) => {
-  const [menuOpen, setMenuOpen] = useState<boolean>(false);
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-
-  const handleClose = () => {
-    setAnchorEl(null);
-    setMenuOpen(false);
-  };
-
-  const handleAnchorMenu = (e: any) => {
-    setAnchorEl(e.currentTarget);
-    setMenuOpen(true);
-  };
-
-  const handleMenuItemClicked = (action: SummarySectionMenuActions) => {
-    onMenuClick?.({ type: action, context: { id: user._id! } });
-    handleClose();
-  };
-
-  if (cardType === 'thread') {
-    // TODO: The text excerpt needs to be truncated if it's too long
-    return (
-      <Box>
-        <Typography
-          sx={{ color: theme.palette.primary.greyDisabled, textAlign: 'left' }}
-          variant='body1'
-        >
-          Excerpt of latest message
-        </Typography>
-      </Box>
-    );
-  }
-
-  // Return info (location and something else?)
-  return (
-    <Box display='flex' width={'100%'}>
-      <Box display='flex' flexGrow={3}>
-        <Box>
-          <LocationPinIcon sx={{ color: theme.palette.primary.greyDisabled }} />
-        </Box>
-        <Box>
-          <Typography variant='body1' sx={{ color: theme.palette.primary.greyDisabled }}>
-            {user?.location?.city}, {user?.location?.country}
-          </Typography>
-        </Box>
-      </Box>
-      <Box flexGrow={1} display='flex' justifyContent={'right'}>
-        <ButtonBase onClick={handleAnchorMenu}>
-          <LinearScaleIcon sx={{ color: theme.palette.primary.charcoal }} />
-        </ButtonBase>
-      </Box>
-      <Menu
-        id='simple-menu'
-        anchorEl={anchorEl}
-        open={menuOpen}
-        onClose={handleClose}
-        MenuListProps={{
-          'aria-labelledby': 'basic-button',
-        }}
-      >
-        <MenuItem onClick={() => handleMenuItemClicked(SummarySectionMenuActions.RemoveUser)}>
-          <Typography sx={{ color: theme.palette.primary.burntOrangeCancelError }}>
-            Remove
-          </Typography>
-        </MenuItem>
-      </Menu>
-    </Box>
-  );
-};
-
-export function NewConversationHeader() {
-  return (
-    <Box p={2}>
-      <Box id='header-with-cancel-enclosure' display='flex' mb={1} justifyContent={'space-between'}>
-        <Box>
-          <Typography
-            sx={{
-              color: theme.palette.primary.charcoal,
-              fontWeight: 'bold',
-              fontSize: '1.25rem',
-              lineHeight: '1.75em',
-            }}
-          >
-            New Conversation
-          </Typography>
-        </Box>
-        <Box>
-          <ButtonBase>
-            <Typography
-              sx={{
-                color: theme.palette.primary.primaryColorDarkBlue,
-                textDecoration: 'underline',
-              }}
-            >
-              Cancel
-            </Typography>
-          </ButtonBase>
-        </Box>
-      </Box>
-      <Box>
-        {/* Search for people box */}
-        <SearchInputField
-          handleSearch={() => {}}
-          keyword=''
-          fullWidth={true}
-          placeholder='Search people'
-        />
-      </Box>
-    </Box>
-  );
-}
 
 // For the messages. This should be rendered in a container that controls how the message displays to the left or to the right.
 export const MessageBlurb = ({
